requirements.txt
matplotlib==3.4.3
numpy==1.20.3
open3d==0.16.0
opencv_python==4.5.5.64
pandas==1.3.4
pkbar==0.5
polarTransform==2.0.0
protobuf==3.19.1
seaborn==0.11.2
Shapely==1.8.4
tensorboard==2.6.0
torch_summary==1.4.5
torchsummary==1.5.1
